Transition from ex-GORM structs to sqlc structs (5/5)
Replace old used-to-be-GORM datastructures (#104305) with sqlc-generated
structs. This also makes it possible to use more specific structs that
are more taylored to the specific queries, increasing efficiency.

This commit deals with the remaining areas, like the job deleter, task
timeout checker, and task state machine. And anything else to get things
running again.

Functional changes are kept to a minimum, as the API still serves the
same data.

Because this work covers so much of Flamenco's code, it's been split up
into different commits. Each commit brings Flamenco to a state where it
compiles and unit tests pass. Only the result of the final commit has
actually been tested properly.

Ref: #104343

+++ internal/manager/persistence/jobs_query_test.go
@@ -49,7 +49,7 @@ func TestQueryJobTaskSummaries(t *testing.T) {
 	assert.Len(t, summaries, len(expectTaskUUIDs))
 	for index, summary := range summaries {
 		assert.True(t, expectTaskUUIDs[summary.UUID], "%q should be in %v", summary.UUID, expectTaskUUIDs)
-		assert.Equal(t, index+1, summary.IndexInJob)
+		assert.Equal(t, int64(index+1), summary.IndexInJob)
 	}
 }
